DRF2Vue/backServer/requirements.txt
absl-py==1.4.0
appnope==0.1.3
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.0
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
comm==0.1.2
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.5
decorator==4.4.2
deepspeech==0.9.3
Django==3.2
django-cors-headers==3.14.0
django-debug-toolbar==4.0.0
django-extensions==3.2.1
django-seed==0.3.1
djangorestframework==3.14.0
entrypoints==0.4
exceptiongroup==1.1.1
executing==1.2.0
Faker==13.3.4
ffmpeg==1.4
ffmpeg-python==0.2.0
flatbuffers==23.3.3
fonttools==4.38.0
future==0.18.3
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.0
h11==0.14.0
h5py==3.8.0
idna==3.4
imageio==2.28.1
imageio-ffmpeg==0.4.8
importlib-metadata==6.6.0
ipykernel==6.20.1
ipython==8.12.0
jax==0.4.8
jedi==0.18.2
Jinja2==3.1.2
jinja2-time==0.2.0
joblib==1.2.0
jupyter_client==7.4.8
jupyter_core==5.1.3
kaggle==1.5.12
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
lxml==4.9.2
make==0.1.6.post2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.6.2
matplotlib-inline==0.1.6
ml-dtypes==0.1.0
moviepy==1.0.3
nest-asyncio==1.5.6
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
outcome==1.2.0
packaging==23.0
pandas==1.5.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==2.6.2
plotly==5.11.0
proglog==0.1.10
prompt-toolkit==3.0.38
protobuf==4.22.4
psutil==5.9.4
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydub==0.25.1
Pygments==2.14.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==7.0.0
pytz==2023.3
pywin32==305
pyzmq==24.0.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.0
scipy==1.10.0
seaborn==0.12.2
selenium==4.8.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
SpeechRecognition==3.10.0
sqlparse==0.4.3
stack-data==0.6.2
tenacity==8.1.0
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
text-unidecode==1.3
threadpoolctl==3.1.0
toposort==1.7
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
webdriver-manager==3.8.6
Werkzeug==2.3.3
wrapt==1.14.1
wsproto==1.2.0
zipp==3.15.0
